+++ src/scmdata/run.py
@@ -32,7 +32,8 @@ from .filters import (
 from .groupby import RunGroupBy
 from .netcdf import inject_nc_methods
 from .offsets import generate_range, to_offset
-from .pyam_compat import Axes, IamDataFrame, LongDatetimeIamDataFrame
+from .plotting import inject_plotting_methods
+from .pyam_compat import IamDataFrame, LongDatetimeIamDataFrame
 from .time import TimePoints
 from .timeseries import TimeSeries
 from .units import UnitConverter
@@ -557,7 +558,7 @@ class ScmRun:  # pylint: disable=too-many-public-methods
         ValueError
             If the metadata are not unique between timeseries
         """
-        df = pd.DataFrame(np.asarray(self._ts))
+        df = pd.DataFrame(self.values)
         _meta = self.meta if meta is None else self.meta[meta]
         if check_duplicated and _meta.duplicated().any():
             raise ValueError("Duplicated meta values")
@@ -585,7 +586,7 @@ class ScmRun:  # pylint: disable=too-many-public-methods
 
         Calls :func:`timeseries`
         """
-        return np.asarray(self._ts)
+        return np.asarray([ts._data.values for ts in self._ts])
 
     @property
     def meta(self) -> pd.DataFrame:
@@ -1475,7 +1476,7 @@ class ScmRun:  # pylint: disable=too-many-public-methods
 
         return df_append([self, other], inplace=inplace, duplicate_msg=duplicate_msg)
 
-    def to_iamdataframe(self) -> LongDatetimeIamDataFrame:
+    def to_iamdataframe(self) -> LongDatetimeIamDataFrame:  # pragma: no cover
         """
         Convert to a :class:`LongDatetimeIamDataFrame` instance.
 
@@ -1509,46 +1510,7 @@ class ScmRun:  # pylint: disable=too-many-public-methods
         fname
             Path to write the file into
         """
-        self.to_iamdataframe().to_csv(fname, **kwargs)
-
-    def line_plot(self, x: str = "time", y: str = "value", **kwargs: Any) -> Axes:
-        """
-        Plot a line chart.
-
-        See :func:`pyam.IamDataFrame.line_plot` for more information.
-        """
-        return self.to_iamdataframe().line_plot(x, y, **kwargs)  # pragma: no cover
-
-    def scatter(self, x: str, y: str, **kwargs: Any) -> Axes:
-        """
-        Plot a scatter chart using metadata columns.
-
-        See :func:`pyam.plotting.scatter` for details.
-        """
-        self.to_iamdataframe().scatter(x, y, **kwargs)  # pragma: no cover
-
-    def region_plot(self, **kwargs: Any) -> Axes:
-        """
-        Plot regional data for a single model, scenario, variable, and year.
-
-        See :class:`pyam.plotting.region_plot` for details.
-        """
-        return self.to_iamdataframe().region_plot(**kwargs)  # pragma: no cover
-
-    def pivot_table(
-        self,
-        index: Union[str, List[str]],
-        columns: Union[str, List[str]],
-        **kwargs: Any,
-    ) -> pd.DataFrame:
-        """
-        Pivot the underlying data series.
-
-        See :func:`pyam.core.IamDataFrame.pivot_table` for details.
-        """
-        return self.to_iamdataframe().pivot_table(
-            index, columns, **kwargs
-        )  # pragma: no cover
+        self.timeseries().reset_index().to_csv(fname, **kwargs)
 
     def reduce(self, func, dim=None, axis=None, **kwargs):
         """
@@ -1734,3 +1696,4 @@ def _handle_potential_duplicates_in_append(data, duplicate_msg):
 
 inject_binary_ops(ScmRun)
 inject_nc_methods(ScmRun)
+inject_plotting_methods(ScmRun)